+++ src/commands/list.js
@@ -3,10 +3,11 @@ const { cli } = require('cli-ux')
 
 class ListCommand extends Command {
   async run () {
-    const { flags } = this.parse(ListCommand)
+    const { args, flags } = this.parse(ListCommand)
 
     const path = require('path')
     const fs = require('fs')
+    const util = require('util')
 
     let dir, file
     if (flags.use) {
@@ -25,37 +26,125 @@ class ListCommand extends Command {
     const data = JSON.parse(fs.readFileSync(file, 'utf-8'))
     if (!data.length) return
 
-    if (flags.json) {
-      this.log(data)
-    } else {
-      cli.table(
-        data,
-        {
-          name: { header: 'Profile name' },
-          host: { header: 'Host or IP' },
-          port: { header: 'Port' },
-          user: { header: 'User name' },
-          pass: { header: 'Password', get: (x) => x.pass || '' },
-          key: { header: 'Key file', get: (x) => x.key || '' },
-          time: { header: 'Timestamp' }
-        },
-        {
-          sort: 'name'
-        }
-      )
+    switch (args.format) {
+      case 'csv':
+        this.log(
+          [
+            'Profile name',
+            'Host or IP',
+            'Port',
+            'User name',
+            'Password',
+            'Key file',
+            'Timestamp'
+          ].join(',')
+        )
+        this.log(
+          data
+            .map(
+              (x) =>
+                x.name +
+                ',' +
+                x.host +
+                ',' +
+                x.port +
+                ',' +
+                x.user +
+                ',' +
+                (x.pass || '') +
+                ',' +
+                (x.key || '') +
+                ',' +
+                x.time
+            )
+            .join('\n')
+        )
+        break
+      case 'json':
+        this.log(
+          util.inspect(data, {
+            colors: true
+          })
+        )
+        break
+      case 'json-minified':
+        this.log(JSON.stringify(data))
+        break
+      case 'table':
+        cli.table(
+          data,
+          {
+            name: { header: 'Profile name' },
+            host: { header: 'Host or IP' },
+            port: { header: 'Port' },
+            user: { header: 'User name' },
+            pass: { header: 'Password', get: (x) => x.pass || '' },
+            key: { header: 'Key file', get: (x) => x.key || '' },
+            time: { header: 'Timestamp' }
+          },
+          {
+            sort: 'name'
+          }
+        )
+        break
+      case 'tsv':
+        this.log(
+          [
+            'Profile name',
+            'Host or IP',
+            'Port',
+            'User name',
+            'Password',
+            'Key file',
+            'Timestamp'
+          ].join('\t')
+        )
+        this.log(
+          data
+            .map(
+              (x) =>
+                x.name +
+                '\t' +
+                x.host +
+                '\t' +
+                x.port +
+                '\t' +
+                x.user +
+                '\t' +
+                (x.pass || '') +
+                '\t' +
+                (x.key || '') +
+                '\t' +
+                x.time
+            )
+            .join('\n')
+        )
+        break
     }
   }
 }
+
+const formats = ['csv', 'json', 'json-minified', 'table', 'tsv']
 
 ListCommand.aliases = ['ls']
 
 ListCommand.description = 'list profiles'
 
+ListCommand.args = [
+  {
+    name: 'format',
+    description: 'format to output profiles in',
+    default: 'table',
+    options: formats
+  }
+]
 ListCommand.flags = {
-  use: flags.string({ description: 'path to custom sshpm configuration file' }),
-  json: flags.boolean({ char: 'j', description: 'output in JSON format' })
+  use: flags.string({ description: 'path to custom sshpm configuration file' })
 }
 
-ListCommand.examples = ['$ sshpm list', '$ sshpm list -j']
+ListCommand.examples = [
+  '$ sshpm list',
+  ...formats.map((x) => '$ sshpm list ' + x)
+]
 
 module.exports = ListCommand